genAI/requirements.txt
fastapi
pydantic
python-dotenv
uvicorn[standard]
pandas
openai
pytesseract
requests
pillow